Works...
No longer have Memory issues so this issue can be put to rest, however,
the runtime is abysmal. It takes 3 min to do monroe county, and I let
it run for 52 min and id didn’t finish NY. Going to need to spend a lot
of time working on that.